Update padding-and-strides_origin.md
I think the formula should change \lfloor and \rfloor to \lceil and \rceil. For example, if an (5x5) input performing convolution operation, and the kernel has size of 3x3, strides is 2, the output should be (2x2), but if use "round down", the result is 1x1, we could get right answer if we use "round up".
And if we use the kernel has same size with input, and no padding, strides is 1, the output is 0x0 if we use "round down",but we know it must be 1x1, "round down" could solve this problem.
If my thought has any wrong, hope you could tell me, thanks.